Zmień typ przechowywania pozycji boida na zmiennoprzecinkowy

+++ boids-seq/Boid.cpp
@@ -1,9 +1,9 @@
 #include "Boid.h"
 
-Boid::Boid(int x, int y){
+Boid::Boid(double x, double y){
   position = Point2D(x,y);
-  int a = rand()%(2*MAX_SPEED)-MAX_SPEED;
-  int b  = rand()%(2*MAX_SPEED)-MAX_SPEED;
+  double a = rand()%(2*MAX_SPEED)-MAX_SPEED;
+  double b  = rand()%(2*MAX_SPEED)-MAX_SPEED;
   velocity = Vector2D(a, b);
   acceleration = Vector2D(0, 0);
   separate_force = Vector2D(0, 0);
